Consolidate api/tests
