+++ api/tests/test_api.py
@@ -44,6 +44,71 @@ class AuthorTests(TestCase):
         res = self.client.get('/api/v1/authors/')
         self.assertEqual(res.status_code, 200)
 
+class PostTests(TestCase):
+    def setUp(self) -> None:
+        self.client = Client()
+        self.user = get_user_model().objects.create_user(username='bob', password='password')
+        self.post = Post.objects.create(
+            title=POST_DATA['title'],
+            description=POST_DATA['description'],
+            content_type=POST_DATA['content_type'],
+            content=POST_DATA['content'],
+            author_id=self.user.id,
+            unlisted=POST_DATA['unlisted'])
+
+        self.other_user = get_user_model().objects.create_user(username='alice', password='password')
+        self.post_by_other_user = Post.objects.create(
+            title=POST_DATA['title'],
+            description=POST_DATA['description'],
+            content_type=POST_DATA['content_type'],
+            content=POST_DATA['content'],
+            author_id=self.other_user.id,
+            unlisted=POST_DATA['unlisted'])
+        self.post_by_other_user.save()
+
+    def test_posts(self):
+        self.client.login(username='bob', password='password')
+        res = self.client.get(f'/api/v1/authors/{self.user.id}/posts/')
+        self.assertEqual(res.status_code, 200)
+        body = json.loads(res.content.decode('utf-8'))
+        self.assertEqual(body['type'], 'posts')
+        self.assertEqual(len(body['items']), 1)
+
+        for post in body['items']:
+            self.assertIn('id', post)
+            self.assertIn('title', post)
+            self.assertIn('content', post)
+            self.assertIn('author', post)
+            self.assertIn('visibility', post)
+            self.assertIn('unlisted', post)
+            self.assertIn('type', post)
+            self.assertIn('contentType', post)
+            self.assertIn('published', post)
+            self.assertIn('url', post)
+            self.assertIn('categories', post)
+
+    def test_posts_require_login(self):
+        res = self.client.get(f'/api/v1/authors/{self.user.id}/posts/')
+        self.assertEqual(res.status_code, 403)
+
+    def test_post_detail(self):
+        self.client.login(username='bob', password='password')
+        res = self.client.get(f'/api/v1/authors/{self.user.id}/posts/{self.post.id}/')
+        self.assertEqual(res.status_code, 200)
+        post = json.loads(res.content.decode('utf-8'))
+
+        self.assertIn('id', post)
+        self.assertIn('title', post)
+        self.assertIn('content', post)
+        self.assertIn('author', post)
+        self.assertIn('visibility', post)
+        self.assertIn('unlisted', post)
+        self.assertIn('type', post)
+        self.assertIn('contentType', post)
+        self.assertIn('published', post)
+        self.assertIn('url', post)
+        self.assertIn('categories', post)
+
 
 class ImageTests(TestCase):
     def setUp(self) -> None:
